Run refiners as long as validators passed (#1068)
* Run refiners as long as validators passed

Previously refiners ran only when there were no past failures during
validation of the value itself or its subvalues. This behavior was
causing inferior UX when validating forms with additional validations
defined as refiners. If there was a refinement error in some subfield,
the refiners defined on the parent of that subfield were not executed.
Only after fixing the error in the subfield did the user receive
information about the failure on the parent field.

This commit changes the condition under which refinements are run.
Refinements are executed if there were no previous failures or those
failures came exclusively from other refiners.

The change helps in the form validation scenario because as long as the
form structure is valid (which is checked using `validator`s), failures
from all refiners are collected.

Refiners still run only if the general structure of the value is
correct. They will get passed a structurally-valid value. The only
difference now is that those values were may not have been checked via a
refiner.

If some check is so important that other refiners should not
run if that check fails, that check should be defined as a `validator`,
not a `refiner`. Failing that check will not run other refiners.

Fixes https://github.com/ianstormtaylor/superstruct/issues/979

* refactor refinement code

Co-authored-by: Ian Storm Taylor <ian@ianstormtaylor.com>