Fix ESM/CJS compatibility (#1)
## Description

Superstruct cannot be imported in a TypeScript project that uses a
`moduleResolution` of `node16` or `nodenext`. Any attempt to do so will
result in type errors such as the following:

```
node_modules/superstruct/dist/index.d.ts:1:15 - error TS2834: Relative import paths need explicit file extensions in ECMAScript imports when '--moduleResolution' is 'node16' or 'nodenext'. Consider adding an extension to the import path.

1 export * from './error';
                ~~~~~~~~~

node_modules/superstruct/dist/index.d.ts:2:15 - error TS2834: Relative import paths need explicit file extensions in ECMAScript imports when '--moduleResolution' is 'node16' or 'nodenext'. Consider adding an extension to the import path.

2 export * from './struct';
                ~~~~~~~~~~

node_modules/superstruct/dist/index.d.ts:3:15 - error TS2834: Relative import paths need explicit file extensions in ECMAScript imports when '--moduleResolution' is 'node16' or 'nodenext'. Consider adding an extension to the import path.

3 export * from './structs/coercions';
                ~~~~~~~~~~~~~~~~~~~~~

...
```

This is happening because although Superstruct is defined as an ESM
library — the `module` field in `package.json` is set to `true` — its
published TypeScript declaration files aren't ESM-compatible. This is
due to the fact that imports are missing extensions, which is a
requirement for ESM.

At the same time, although Rollup is supposedly configured to emit both
ESM- and CJS-compatible files, TypeScript does not know how to use the
CJS variants. The `exports` field is the way to instruct TypeScript on
how to do this, but `package.json` is missing this field. This makes it
impossible to use Superstruct in a non-ESM library.

This commit fixes both of these issues by making the following changes:

- All imports in TypeScript files now use an extension of `.js`. This
may seem like an odd choice, as there are no JavaScript files in the
`src/` directory, but TypeScript doesn't resolve a module by trying to
find a file with the given extension; it tries to find a declaration
file that maps to the module path, based on a set of rules. Therefore,
the file extension is really just a formality. In the end, Rollup is
consolidating implementation files into one, so all of the imports will
go away.
- Type declaration files have been split into ESM and CJS variants.
Since Rollup generates ESM variants with an `.mjs` extension and CJS
variants with a `.cjs` extension, imports in declaration files also use
either `.mjs` or `.cjs` depending on the variant.
- `package.json` now has an `exports` field which instructs TypeScript
how to resolve modules from either an ESM or CJS perspective.

In addition, the tests have been updated to use Vitest instead of Babel
so that they can read directly from the TypeScript configuration instead
of having to use an explicit transform step.

## References

This commit is derived from the following PR on the Superstruct repo:
<https://github.com/ianstormtaylor/superstruct/pull/1211>

For more information on how TypeScript resolves modules, the "Modules
Reference" section of the TypeScript handbook is quite helpful, and in
particular these sections:

-
<https://www.typescriptlang.org/docs/handbook/modules/theory.html#the-role-of-declaration-files>
-
<https://www.typescriptlang.org/docs/handbook/modules/reference.html#node16-nodenext>